feat: complete threat space search: available to detect vectory by continuous 3-threats; perf: improved minimax performence

+++ src/botzone.c
@@ -24,7 +24,7 @@ void print_pos(point_t p)
 
 int main()
 {
-    const int tim = record_time();
+    const double tim = record_time();
     log_lock();
     init();
